🚚mv (option) seperate option from main library
